Review-Fixes + Trainings-Seed: Shift-Augmentierung, selektiver Sync, Changelog (1.4.5)
- OCR: Shift-Augmentierung (+-1px) — Ziffern generalisieren ueber alle
  Boxen (1->7-Rollover-Bug); Modell float16 (8.7MB)
- nuc_feedback_sync: Zeitstempel-Logs, nur Disagreements/Events + jedes
  20. Routine-Sample, Prune erst nach erfolgreichem Push, robustere
  pull/push-Fehlerbehandlung
- Deploy-Key: nur noch git_deploy_key_base64 (config + run.sh)
- CHANGELOG.md im HA-Format
- training-data/ mit kompletter kuratierter Historie geseedet (571 Samples)
  — Training laeuft ueberall gegen die geteilte Git-Wahrheit

+++ scripts/ocr/local_reader.py
@@ -22,7 +22,7 @@ MODEL_FILE = Path(__file__).with_name("model.npz")
 class LocalReader:
     def __init__(self, model_file: Path = MODEL_FILE, k: int = 3):
         m = np.load(model_file, allow_pickle=False)
-        self.X = m["X"]
+        self.X = m["X"].astype(np.float32)  # ggf. float16-komprimiert gespeichert
         self.y = m["y"]
         self.slots = m["slots"] if "slots" in m.files else None
         self.k = k
